requirements.txt
clip==1.0
easydict==1.13
numpy==2.3.5
openai==2.8.1
pandas==2.3.3
Pillow==12.0.0
PyYAML==6.0.2
PyYAML==6.0.3
scipy==1.16.3
torch==2.6.0
tqdm==4.67.1
transformers==4.56.2
